+++ app/dashboard/karbooms/partners-list/_components/partners-list-item-component.tsx
@@ -11,6 +11,8 @@ import useDeletePartnerEndpoint from "../_hooks/use-delete-partner-endpoint";
 
 import { ACTIVITY_STATUS_FA } from "../../_constants/activity-status-fa";
 import { ACTIVITY_STATUS_COLORS } from "../../_constants/activity-status-colors";
+import { useKarboomsStore } from "../../_providers/karbooms-store-provider";
+import { useSnackbar } from "notistack";
 
 export default function PartnersListItemComponent({
   item,
@@ -18,9 +20,14 @@ export default function PartnersListItemComponent({
   onEdit,
 }: PartnersListItemProps) {
   const { full_name, phone, avatar, share, status } = item;
-  const { mutate: deletePartner } = useDeletePartnerEndpoint();
+
+  const { enqueueSnackbar } = useSnackbar();
 
   const [menuAnchor, setMenuAnchor] = useState<HTMLElement | null>(null);
+
+  const karboomRoles = useKarboomsStore((state) => state.roles);
+
+  const { mutate: deletePartner } = useDeletePartnerEndpoint();
 
   const handleOpenMenu = (event: MouseEvent<HTMLButtonElement>) => {
     setMenuAnchor(event.currentTarget);
@@ -36,8 +43,14 @@ export default function PartnersListItemComponent({
   };
 
   const handleDelete = () => {
-    handleCloseMenu();
-    deletePartner(item.id);
+    if (karboomRoles.includes("owner")) {
+      handleCloseMenu();
+      deletePartner(item.id);
+    } else
+      enqueueSnackbar({
+        variant: "warning",
+        message: "فقط سازنده کاربوم میتواند مالک حذف کند",
+      });
   };
 
   return (
